Move most implementation into pkg
