+++ master/internal/core_checkpoint.go
@@ -1,25 +1,18 @@
 package internal
 
 import (
-	"archive/tar"
-	"archive/zip"
 	"bufio"
-	"compress/gzip"
 	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
-	"strings"
 
-	"github.com/hashicorp/go-multierror"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
+
+	"github.com/determined-ai/determined/master/pkg/checkpoints"
+	"github.com/determined-ai/determined/master/pkg/checkpoints/archive"
 
 	"github.com/determined-ai/determined/master/internal/api"
 	"github.com/determined-ai/determined/master/pkg/ptrs"
@@ -33,95 +26,21 @@ const (
 	MIMEApplicationZip = "application/zip"
 )
 
-func storageConfig2Str(config any) string {
-	switch config.(type) {
-	case expconf.AzureConfig:
-		return "azure"
-	case expconf.GCSConfig:
-		return "gcs"
-	case expconf.HDFSConfig:
-		return "hdfs"
-	case expconf.S3Config:
-		return "s3"
-	case expconf.SharedFSConfig:
-		return "shared_fs"
+func mimeToArchiveType(mimeType string) archive.ArchiveType {
+	switch mimeType {
+	case MIMEApplicationGZip:
+		return archive.ArchiveTgz
+	case MIMEApplicationZip:
+		return archive.ArchiveZip
 	default:
-		return "unknown"
+		return archive.ArchiveUnknown
 	}
 }
 
-type archiveWriter interface {
-	WriteHeader(path string, size int64) error
-	Write(b []byte) (int, error)
-	Close() error
-}
-
-type archiveClosers struct {
-	closers []io.Closer
-}
-
-// Close() closes all items in closers in reverse order.
-func (ac *archiveClosers) Close() error {
-	for i := len(ac.closers) - 1; i >= 0; i-- {
-		err := ac.closers[i].Close()
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-type tarArchiveWriter struct {
-	archiveClosers
-	tw *tar.Writer
-}
-
-func (aw *tarArchiveWriter) WriteHeader(path string, size int64) error {
-	hdr := tar.Header{
-		Name: path,
-		Mode: 0666,
-		Size: size,
-	}
-	if strings.HasSuffix(path, "/") {
-		// This a directory
-		hdr.Mode = 0777
-	}
-	return aw.tw.WriteHeader(&hdr)
-}
-
-func (aw *tarArchiveWriter) Write(p []byte) (int, error) {
-	return aw.tw.Write(p)
-}
-
-type zipArchiveWriter struct {
-	archiveClosers
-	zw        *zip.Writer
-	zwContent io.Writer
-}
-
-func (aw *zipArchiveWriter) WriteHeader(path string, size int64) error {
-	// Zip by default sets mode 0666 and 0777 for files and folders respectively
-	zwc, err := aw.zw.Create(path)
-	if err != nil {
-		return err
-	}
-	aw.zwContent = zwc
-	return nil
-}
-
-func (aw *zipArchiveWriter) Write(p []byte) (int, error) {
-	var w io.Writer
-	// Guard against the mistake where WriteHeader() is not called before
-	// calling Write(). The AWS SDK likely will not make this mistake but
-	// zipArchiveWriter is not just limited to being used with AWS.
-	if aw.zwContent == nil {
-		return 0, nil
-	}
-	w = aw.zwContent
-	return w.Write(p)
-}
-
-// delayWriter wraps bufio.Writer and add Close().
+// Since Echo does not send an http status code until the first write to the ResponseWriter.
+// We use delayWriter to buffer our writes, which effectively delays sending the status code
+// until we are more confident the download will succeed. delayWriter wraps bufio.Writer
+// and adds Close().
 type delayWriter struct {
 	next *bufio.Writer
 }
@@ -138,223 +57,6 @@ func (w *delayWriter) Close() error {
 func newDelayWriter(w io.Writer, delayBytes int) *delayWriter {
 	return &delayWriter{
 		next: bufio.NewWriterSize(w, delayBytes),
-	}
-}
-
-// seqWriterAt satisfies S3 APIs' io.WriterAt interface while staying sequential.
-// To use it with s3manager.Downloader, its concurrency needs be set to 1.
-// Ref: https://docs.aws.amazon.com/sdk-for-go/api/service/s3/s3manager/#Downloader
-type seqWriterAt struct {
-	next    io.Writer
-	written int64
-}
-
-func newSeqWriterAt(w io.Writer) *seqWriterAt {
-	return &seqWriterAt{next: w}
-}
-
-// WriteAt writes the content in buffer p.
-func (w *seqWriterAt) WriteAt(p []byte, off int64) (int, error) {
-	if off != w.written {
-		return 0, fmt.Errorf(
-			"only supporting sequential writes,"+
-				" writing at offset %d while %d bytes have been written",
-			off, w.written)
-	}
-	n, err := w.next.Write(p)
-	w.written += int64(n)
-	if err != nil {
-		return 0, err
-	}
-
-	return n, err
-}
-
-// BatchDownloadIterator implements s3's BatchDownloadIterator API.
-type batchDownloadIterator struct {
-	// The objects we are writing
-	objects []*s3.Object
-	// The output we are writing to
-	aw archiveWriter
-	// Internal states
-	err    error
-	pos    int
-	bucket string
-	prefix string
-}
-
-// Next() returns true if the next item is available.
-func (i *batchDownloadIterator) Next() bool {
-	i.pos++
-	if i.pos == len(i.objects) {
-		return false
-	}
-	pathname := strings.TrimPrefix(*i.objects[i.pos].Key, i.prefix)
-	err := i.aw.WriteHeader(pathname, *i.objects[i.pos].Size)
-	if err != nil {
-		i.err = err
-		return false
-	}
-	return true
-}
-
-// Err() eturns the error if any.
-func (i *batchDownloadIterator) Err() error {
-	return i.err
-}
-
-// DownloadObject() eturns a DownloadObject.
-func (i *batchDownloadIterator) DownloadObject() s3manager.BatchDownloadObject {
-	return s3manager.BatchDownloadObject{
-		Object: &s3.GetObjectInput{
-			Bucket: &i.bucket,
-			Key:    i.objects[i.pos].Key,
-		},
-		Writer: newSeqWriterAt(i.aw),
-	}
-}
-
-func newBatchDownloadIterator(aw archiveWriter,
-	bucket string, prefix string, objs []*s3.Object) *batchDownloadIterator {
-	if !strings.HasSuffix(prefix, "/") {
-		prefix += "/"
-	}
-	return &batchDownloadIterator{
-		aw:      aw,
-		bucket:  bucket,
-		prefix:  prefix,
-		objects: objs,
-		pos:     -1,
-	}
-}
-
-func getS3BucketRegion(ctx context.Context, bucket string) (string, error) {
-	sess, err := session.NewSession(&aws.Config{
-		Region: aws.String("us-west-2"),
-	})
-	if err != nil {
-		return "", nil
-	}
-
-	out, err := s3.New(sess).GetBucketLocationWithContext(ctx, &s3.GetBucketLocationInput{
-		Bucket: &bucket,
-	})
-	if err != nil {
-		return "", err
-	}
-
-	return *out.LocationConstraint, nil
-}
-
-type checkpointDownloader interface {
-	download(ctx context.Context) error
-}
-
-type s3Downloader struct {
-	aw     archiveWriter
-	bucket string
-	prefix string
-}
-
-func (d *s3Downloader) download(ctx context.Context) error {
-	region, err := getS3BucketRegion(ctx, d.bucket)
-	if err != nil {
-		return err
-	}
-	sess, err := session.NewSession(&aws.Config{
-		Region: &region,
-	})
-	if err != nil {
-		return err
-	}
-	// We do not pass in credentials explicitly. Instead, we reply on
-	// the existing AWS credentials.
-	s3client := s3.New(sess)
-
-	var merr error
-	downloader := s3manager.NewDownloader(sess, func(d *s3manager.Downloader) {
-		d.Concurrency = 1 // Setting concurrency to 1 to use seqWriterAt
-	})
-	funcReadPage := func(output *s3.ListObjectsV2Output, lastPage bool) bool {
-		iter := newBatchDownloadIterator(d.aw, d.bucket, d.prefix, output.Contents)
-		// Download every bucket in this page
-		err = downloader.DownloadWithIterator(ctx, iter)
-		if iter.Err() != nil {
-			merr = multierror.Append(merr, iter.Err())
-		}
-		if err != nil {
-			merr = multierror.Append(merr, err)
-		}
-
-		// Return False to stop paging
-		return merr == nil
-	}
-	err = s3client.ListObjectsV2PagesWithContext(
-		ctx,
-		&s3.ListObjectsV2Input{
-			Bucket: &d.bucket,
-			Prefix: &d.prefix,
-		},
-		funcReadPage,
-	)
-	if err != nil {
-		merr = multierror.Append(merr, err)
-	}
-	if merr != nil {
-		return fmt.Errorf("checkpoint download failed: %w", merr)
-	}
-	return nil
-}
-
-func newDownloader(
-	storageConfig *expconf.CheckpointStorageConfig,
-	aw archiveWriter,
-	id string,
-) (checkpointDownloader, error) {
-	switch storage := storageConfig.GetUnionMember().(type) {
-	case expconf.S3Config:
-		prefix := ""
-		if storage.Prefix() != nil {
-			prefix = *storage.Prefix()
-		}
-		return &s3Downloader{
-			aw:     aw,
-			bucket: storage.Bucket(),
-			prefix: strings.TrimLeft(prefix+"/"+id, "/"),
-		}, nil
-	default:
-		return nil, echo.NewHTTPError(http.StatusNotImplemented,
-			fmt.Sprintf("checkpoint download via master is only supported on S3"+
-				", but the checkpoint's storage type is %s", storageConfig2Str(storage)))
-	}
-}
-
-// It is assumed that a http status code is not sent until the first write to w.
-func buildWriterPipeline(w io.Writer, mimeType string) (archiveWriter, error) {
-	// DelayWriter delays the first write until we have successfully downloaded
-	// some bytes and are more confident that the download will succeed.
-	dw := newDelayWriter(w, 16*1024)
-	closers := []io.Closer{dw}
-	switch mimeType {
-	case MIMEApplicationGZip:
-		gz := gzip.NewWriter(dw)
-		closers = append(closers, gz)
-
-		tw := tar.NewWriter(gz)
-		closers = append(closers, tw)
-
-		return &tarArchiveWriter{archiveClosers{closers}, tw}, nil
-
-	case MIMEApplicationZip:
-		zw := zip.NewWriter(dw)
-		closers = append(closers, zw)
-
-		return &zipArchiveWriter{archiveClosers{closers}, zw, nil}, nil
-
-	default:
-		return nil, fmt.Errorf(
-			"MIME type must be %s or %s but got %s",
-			MIMEApplicationGZip, MIMEApplicationZip, mimeType)
 	}
 }
 
@@ -392,24 +94,25 @@ func (m *Master) getCheckpointImpl(
 			fmt.Sprintf("checkpoint %s does not exist", id.String()))
 	}
 
-	writerPipe, err := buildWriterPipeline(content, mimeType)
+	// DelayWriter delays the first write until we have successfully downloaded
+	// some bytes and are more confident that the download will succeed.
+	dw := newDelayWriter(content, 16*1024)
+	downloader, err := checkpoints.NewDownloader(
+		dw, id.String(), storageConfig, mimeToArchiveType(mimeType))
 	if err != nil {
 		return echo.NewHTTPError(http.StatusBadRequest, err.Error())
 	}
 	defer func() {
-		if err := writerPipe.Close(); err != nil && retErr == nil {
-			// Only surface this error if there has been none.
-			retErr = echo.NewHTTPError(http.StatusInternalServerError,
-				fmt.Sprintf("failed to complete checkpoint download: %s", err.Error()))
+		for _, v := range []io.Closer{downloader, dw} {
+			if innerErr := v.Close(); innerErr != nil && retErr == nil {
+				// Only surface this error if there has been none.
+				retErr = echo.NewHTTPError(http.StatusInternalServerError,
+					fmt.Sprintf("failed to complete checkpoint download: %s", innerErr.Error()))
+			}
 		}
 	}()
 
-	downloader, err := newDownloader(storageConfig, writerPipe, id.String())
-	if err != nil {
-		return err
-	}
-
-	err = downloader.download(ctx)
+	err = downloader.Download(ctx)
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError,
 			fmt.Sprintf("unable to download checkpoint %s: %s", id.String(), err.Error()))
@@ -430,7 +133,8 @@ func (m *Master) getCheckpointImpl(
 func (m *Master) getCheckpoint(c echo.Context) error {
 	// Get the MIME type. Only a single type is accepted.
 	mimeType := c.Request().Header.Get("Accept")
-	if mimeType != MIMEApplicationGZip && mimeType != MIMEApplicationZip {
+	if mimeType != MIMEApplicationGZip &&
+		mimeType != MIMEApplicationZip {
 		return echo.NewHTTPError(http.StatusUnsupportedMediaType,
 			fmt.Sprintf("unsupported media type to download a checkpoint: '%s'", mimeType))
 	}
