+++ master/pkg/checkpoints/checkpoints.go
@@ -1,0 +1,67 @@
+package checkpoints
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"strings"
+
+	"github.com/determined-ai/determined/master/pkg/checkpoints/archive"
+	"github.com/determined-ai/determined/master/pkg/checkpoints/s3"
+	"github.com/determined-ai/determined/master/pkg/schemas/expconf"
+)
+
+// CheckpointDownloader defines the interface for downloading checkpoints.
+type CheckpointDownloader interface {
+	Download(ctx context.Context) error
+	Close() error
+}
+
+func storageConfig2Str(config any) string {
+	switch config.(type) {
+	case expconf.AzureConfig:
+		return "azure"
+	case expconf.GCSConfig:
+		return "gcs"
+	case expconf.HDFSConfig:
+		return "hdfs"
+	case expconf.S3Config:
+		return "s3"
+	case expconf.SharedFSConfig:
+		return "shared_fs"
+	default:
+		return "unknown"
+	}
+}
+
+// NewDownloader returns a new CheckpointDownloader that writes to w.
+//
+// - w: the underlying Writer that CheckpointDownloader writes to
+// - id: the UUID string of the checkpoint to be downloaded
+// - storageConfig: the CheckpointStorageConfig
+// - archiveType: The ArchiveType (file format) in which the checkpoint shall
+//                be downloaded
+func NewDownloader(
+	w io.Writer,
+	id string,
+	storageConfig *expconf.CheckpointStorageConfig,
+	archiveType archive.ArchiveType,
+) (CheckpointDownloader, error) {
+	aw, err := archive.NewArchiveWriter(w, archiveType)
+	if err != nil {
+		return nil, err
+	}
+
+	switch storage := storageConfig.GetUnionMember().(type) {
+	case expconf.S3Config:
+		prefix := ""
+		if storage.Prefix() != nil {
+			prefix = *storage.Prefix()
+		}
+		return s3.NewS3Downloader(
+			aw, storage.Bucket(), strings.TrimLeft(prefix+"/"+id, "/")), nil
+	default:
+		return nil, fmt.Errorf("checkpoint download via master is only supported on S3"+
+			", but the checkpoint's storage type is %s", storageConfig2Str(storage))
+	}
+}
